[IMP] point_of_sale: pos_*: adding hoot tests
*: pos_appointment, pos_iot, pos_restaurant_appointment, pos_tyro,
pos_urban_piper, pos_urban_piper_enhancements

Add a base for running boot tests in PoS.

All models available in the PoS have been added via the
`models.ServerModel` tool so that they are available in the `MockServer`

In this way, we can simulate the generation of the model definitions
required for the related model, JS classes and data service.

We can now instantiate the related models and call all the methods
specific to each model.

closes odoo/enterprise#89346

X-original-commit: 1009083986e1bd0ea1c036759a1da0242d8f0700
Related: odoo/odoo#217234
Signed-off-by: David Monnom (moda) <moda@odoo.com>