[FIX] hr_payroll_holidays: time off deferring and public holiday
The aim of this commit is to prevent a sick day or any other type of
time off taken on a public holiday to be deferred.

To reproduce:
- Create a public Holiday for previous month
- Regenerate the work entries for that month
- Compute a batch of payslip for that month, validate and mark it as paid
--> The public holiday should be on the payslip
- OPTIONAL: run the cron `Payroll: Generate pdfs` to create the payslip
  pdf in document and make it available to the concerned employee.
- With the concerned employee, put a time off on the whole week of the
  that month. It should overlap with the public holiday.
- Validate the time off. This should mark the leave as needing to be
  deferred.
- Defer it for next payslip
- Compute a batch of payslip for the following month, validate and mark it as paid

Before this commit:
The overlapping sick day gets deferred completely ignoring the fact it
was a public holiday.

After this commit:
The overlapping sick day doesn't get deferred.

Side note:
Solving this bug, I was able to create another bug going un-noticed by
the CI so I added a test case for it.

opw-4903546

closes odoo/enterprise#91950

X-original-commit: b2d5c55334943aae1e3e1367a898e00d3997ec0a
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>