[FIX] {website_}knowledge: fix portal share panel of an article
Steps to reproduce
===============
As an admin
----------------
1. Go to knowledge
2. Open any article.
3. Add a portal user as a member with write access.

As a portal user
--------------------
1. Open the shared article.
2. On the top right corner, click on members button
--> List of members along with their rights will be shown.

The portal user can open the dropdown to alter the rights of other members.
Though doing so will raise an AccessError (in the logs).

After this commit, the portal user will *no* longer see the
1. Toggle to share article to the web.
2. Dropdown to modify members' rights.

Also, the existing test didn't fail because the `_compute_user_can_write` in
the mocked model `KnowledgeArticle` returned true only when user can write
and user is internal user. As we can have portal user as a writer, the
condition should no longer be applied.

Task-4863456

closes odoo/enterprise#94183

X-original-commit: 9d70e19a511dce5ea13f79994fc688303377eb14
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>