[FIX] website_helpdesk: change "description"
 field behavior

## Short functional explanation of the issue
When a ticket is created and there are no additional fields, the only field
in the ticket is the description field. Moreover, it is not translated.

## Reproduction Steps
1. Go to the general settings. Add a language where the term "description"
differs from the english word (for example, Spanish). Do the same for the
website settings.
2. Go to the website app. Scroll down and change the language. Then,
 click on the translated "contact us" tab. Fill the form and send the
 ticket.
3. Go to the helpdesk app and open the ticket you just created.

### Expected behavior
The description field shouldn't show, as it appears in the description page.
The field should only show if there are additional fields on the "contact
us" page.

### Unexpected behavior
The description field shows, untranslated.

## Origin of the issue
In the ticket code, the field.name is used, instead of the field
.field_description, which is translatable. Moreover, there isn't a condition
checking if the field should appear or not.

opw-4876726

closes odoo/enterprise#90983

X-original-commit: e80648de3021492c80bea3dbea5d96fe8d68fca4
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>