[FIX] hr: close old contract upon creating a new one
When you create a new contract through the salary configurator flow, it will create automatically a new version with the date encoded in the offer wizard. However, the 'contract_end_date' for the old verison is not automatically updated.
The logic has been adjusted to end old contract  date  the day before the new contract's start date.

closes odoo/enterprise#89820

X-original-commit: 09704c95fe5526ea1b635d544c9e111fb8198ef0
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>