[IMP] *: add bottom sheet
*: knowledge,test_spreadsheet_edition,web_studio_ai_fields

A BottomSheet is a UI component in mobile apps that slides up from the
bottom of the screen to display additional content or options. It can be
partially or fully expanded and is commonly used for actions, menus, or
forms. BottomSheets provide a non-intrusive way to show relevant
information without leaving the current screen.

This commit fixes JS tests

task-4551522

closes odoo/enterprise#89119

Related: odoo/odoo#216854
Signed-off-by: Pierre Paridans (app) <app@odoo.com>