[IMP] {website_}appointment: always enable operator step + ui imps
In order to clarify the operator step activation, we do not
make its presence depend on the avatars_display parameter.

Instead, we always show the step if we select user / resource
first for that appointment, and we can at least choose between
two or more users/resources.

When not displaying avatars, we instead show placeholders images,
with a lower height (ratio) as no details are needed. The rest
of the flow is unchanged.

We also take advantage of this opportunity to improve the resource
selection screen for smaller screens, that had images way too large.
Instead, we match what we are doing for users: small image on the
left, content of the card on the right.

A test is moved to ensure it is only tested with appointment, but
not website_appointment, since the rendered screen would not be
the same (slot selection vs user selection). An inlined utility
method was also moved to the class while moving the test.

Task-4898893

Part-of: odoo/enterprise#88637
Related: odoo/upgrade#8014
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>