[FIX] spreadsheet_dashboard_*: correct filter type
The generated domain is wrong when filtering on the date.
It doesn't account for the user timezone because the field matching
is given as a "date" field instead of being a "datetime".

The generated domain looks like `[("create_date", ">=", "2025-06-27")]`
instead of `[("create_date", ">=", "2025-06-27 21:59:59")]`

I'm fixing this in 18.0 because this fix won't affect existing databases
(without updating the modules). The fix would be useless and I don't want
to go though the pain of forward-ports for nothing.
New databases are created in 18.0 every day (latest LTS)

closes odoo/enterprise#89061

Task: 4903362
X-original-commit: f088e60a559f4e39162a7bbfb9b8fc8e6372f859
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>