[FIX] delivery_easypost: include reference on delivery labels
Issue
-----
When using the Easypost delivery service, we don't pass the move reference to
the carrier. This makes it hard for the user to match the deliveries found on
the carrier dashboard with the corresponding Odoo move.

Steps to reproduce
-----
- Set an Easypost delivery method up
- Create a sale
- Add delivery product & get a rate
- Confirm the sale
- Go to the linked delivery & confirm it

--> The printed label has no reference field

Discussion
-----
As per EasyPost:

The reference field within the shipment body is used for transmissions of
electronic data between the shipper and EasyPost, and typically does not get
expressed anywhere on the physical label. To get data to the label, you need to
leverage the "print_custom" fields outlined in the options section of the
shipment.

See also
https://docs.easypost.com/docs/shipments/options

The main reason for the change is to get the info through to the carrier so that
it can get displayed on the user's dashboard.

-----
Ticket:
opw-4681615

closes odoo/enterprise#89332

X-original-commit: de62d0439823b10bc21d163b2f390b9394c68232
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: David Van Droogenbroeck (drod) <drod@odoo.com>