[IMP] account_reports: Add audit checklist and working file support for accounting returns
This commit enhances the functionality of accounting returns, specifically for audit purposes.
It introduces the concept of "Working Files", which are a type of accounting return designed to manually audit the financial and accounting status of a company. For returns of this type, a new view is available. To access it, create a working file, open the corresponding audit, click the top-right button to reveal the top bar, and select the "Balances" view.

The Balances view shows account balances over the audit period, along with additional columns such as:

The difference with the previous year
The audit status, which can be updated directly from the view
This status is also available and editable on some accounting reports like the trial balance and balance sheet.

Users can now create new return types (either tax return or audit) and configure fields such as periodicity and start date.

Additionally, a new model account.return.check.template is introduced, allowing users to define and manage checks (either on audit or tax return). Each check can be:

A manual validation
A document upload
An automated validation based on a domain

Audit checks are now grouped into cycles, which categorize checks by areas like fixed assets, inventory, operating expenses, etc.

Note: While not recommended, it is now technically possible to configure multiple returns of the same type within the same date range.

task-4840028

closes odoo/enterprise#88360

Related: odoo/odoo#220314
Related: odoo/upgrade#8149
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>