[FIX] l10n_mx_edi: commit `l10n_mx_edi_post_time` to DB before sending
Background

In 006e3017037689 we changed the behaviour of `l10n_mx_edi_post_time` so
that it gets set during the CFDI sending process, rather than when the
invoice is posted.

This was needed to ensure that the post time both reflects the sending
time and is re-used if the invoice needs to be re-sent.

Bug description

Since then, several Mexican customers have reported that some CFDIs are
sent duplicated, with a difference of a few seconds in the CFDI's Fecha.

Analysis
A disconnection might occur during sending of the CFDI. In that case,
the PAC will process the CFDI, but in Odoo the transaction will be
rolled back, and the `l10n_mx_edi_post_time` will be lost. When Odoo
re-sends the CFDI, it will take a new `l10n_mx_edi_post_time`.

Solution
Commit before performing the API call so that the
`l10n_mx_edi_post_time` gets committed to DB.

opw-4780096

closes odoo/enterprise#89472

X-original-commit: 062f98ea1f2df7cff45fc0a9a88abcd99d0adc1e
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>