[FIX] stock_barcode: scan package after products
Issue
=====
When a package is scanned as the destination package, if a source
location was previously scanned, the source location will be updated for
every product who will be packed.

How to reproduce
================
- Enable multi-locations and package;
- Create an empty package, two locations and two products;
- Create a delivery in the Barcode app;
- Scan the first location then the first product;
- Scan the second location then the second product;
- Scan the empty package -> The package is rightly assigned as the
  result package for both lines, but the source location of the first
  product was update for the last scanned source location.

Cause of the issue
==================
When a source location was previously scanned, when a line is updated
(`updateLine`), we update the line's source location. Usually, that's
the wanted behavior but in this case, we don't want to the source
location of already processed lines when we scan a destination package.

Solution
========
When calling `updateLine` from `_assignEmptyPackage`, give a key in the
parameters to not update the source location.

opw-4859851

closes odoo/enterprise#90962

X-original-commit: a660976daf85a79dc994cad6a700e1dac2e459e4
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>