[FIX] account_accountant: reco model delete after apply
Before this commit when applying a reco model, the reco model is on the lines
of the move except the liquidity line which means that when we delete the lines
the reco models is lost.
Now when deleting a line, if the statement line had a reco model, we place it
back on the suspense.

Also correcting some demo data where they added the reconcile model on the
liquidity but since the reco model will be added at the creation of the reco
model, that's not needed.

closes odoo/enterprise#90303

Task: 4909214
X-original-commit: 1f0296fd8f999a97a0952322b96041593c0a4dde
Related: odoo/odoo#219204
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>