[FIX] voip: fix cursor displacement in keypad input on mobile.
This commit fixes a bug that caused the cursor in the keypad input to be
displaced when typing on mobile. So, the first entered number was always
to the end. This happens because the `on-touchend` event was calling
another event (click) which may involve some changes in the UI that
caused this.

Steps to produce the bug:
1. Open VoIP wizard from mobile.
2. try to add numbers to the keypad.
3. You will see that the first number is always at the end.

closes odoo/enterprise#89232

X-original-commit: 668467678312632b5b9ecf234280690151bf3078
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>