[FIX] account_avatax: stop calling the wrong super hook
This is a typo from the most recent refactor [1]. The commit hook is
specific to US avatax, there is no corresponding
super().commit_external_taxes so this can be removed all together.

This didn't break any tests because (at least with the standard code),
super()._uncommit_external_taxes is an empty method that does nothing.

[1] https://github.com/odoo/enterprise/pull/82623

closes odoo/enterprise#89293

X-original-commit: 0865d539944d1beb6f8dd0c697f8a2df097dab7f
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>