[FIX] social_facebook: remove `page_content_activity`
Bug
===
`page_content_activity` has been deprecated, so the API return an error
and the other statistics are set to zero.

https://developers.facebook.com/docs/pages-api/changelog

Now, the only equivalent is `post_activity_by_action_type`, but it
needs to be called on all posts, and it does not support periods
(so even if we call it for each post, we won't be able to compute
the trend). So we simply removed the `stories` statistics for Facebook.

https://developers.facebook.com/docs/graph-api/reference/v23.0/insights

Task-4914555

closes odoo/enterprise#90290

X-original-commit: 35056c5d3646443405c4e6fd5d4b5bd20a6cd42a
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>