[IMP] web_studio: adapt test
In https://github.com/odoo/odoo/pull/216638, the lists of operators
available for selection in the tree editor haver been changed for several
field types. Here we adapt a studio test to that change and make it less
likely to fail if similar changes would be brought.

Task ID: 4894524

closes odoo/enterprise#89233

Related: odoo/odoo#216638
Signed-off-by: Michaël Mattiello (mcm) <mcm@odoo.com>