[FIX] account_accountant: cron company
Before this commit when the cron was triggered in a multi currency environment,
the self.env.company could be a different one from the one of the statement line
that would cause problem when computing the suspense amount since the
currency_rate would be different.

In this example:
- 2 companies one in EUR, one in USD
- On the USD company do a transaction
- the cron is trigger with the self.env.company as the EUR one
- It will do a partial reconcile with an invoice in EUR
- The amount currency of the suspense will be computed by doing the balance *
the currency rate. Since the company is the EUR one, the rate is not one and the
 open currency amount is wrongly computed
- This leads to an error of unbalanced move

no task id

closes odoo/enterprise#90107

X-original-commit: 7d496d1ff2af868fcab2a5911ab3c0dec585acb9
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>