[FIX] account_accountant: Fix full match
Actually, we are matching to blindly for full matches. We only check if aml.move_name or aml.ref is present in st_line label.

But this cause issues with move names like dates, where you could have :
    - move_name = 2025/123
    - label contains 2025/1234567
In this case, we don't want to match, as 2025/123 is not perfectly present in the label.

So this commit update the condition so we need now to have a perfect full match between st_line label and aml.ref or aml.move_name.

no-task

closes odoo/enterprise#89986

X-original-commit: 7b53254c62cad7f890c22467400207c81ebb8895
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>