[IMP] account_reports: Replace annotations by chatter
We are removing the annotations and add them in the chatter instead.
We know which messages are annotations based on the
account_reports_annotation_date.

Since annotations needs to know the lines of the report for them to
know which accounts, moves, ... are in the report, we are also only
fetching the annotations for the line unfolded and not fetching when
an annotation is added or removed.

However this require to iterate through all the lines of a report
and filter empty message that are not empty due to them having some
html tags in them, so it should be a bit slower than previously.

task-4900403

closes odoo/enterprise#89292

Related: odoo/odoo#217168
Related: odoo/upgrade#7994
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>