[IMP] hr: Improving version list view.
In this PR, we introduced some improvement in the version list view. Mainly: making the list row clickable and it will open the employee form with the right view, Adding the employee to the search view in the first position, and enabling the versions menuitem in the payroll app (no need for debug mode to display it).
Related task: 4902608.

closes odoo/enterprise#89284

Related: odoo/odoo#217146
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>