[IMP] {website_}appointment: inline and simplify kanban record menu
In order to make the use of the kanban view more snappy, we remove
completely the 'Action' record on the Kanban Records. Instead, we
inline on the card two of the four actions: Share, and Edit (now
renamed to 'Configure')

The two remaining options have dedicated buttons (no need to open the
small dropdown menu to use them). The design is carefully handled
to make it as responsive as before when it comes to various screen
sizes. (A small imp is also done to avoid strange display in grouped
kanban helper template cards)

'Configure' simply opens the form view, where removed actions from
the kanban menu can still be used.

'Share' is also made simpler, and comes with a new component
AppointmentQuickShareButton. Instead of opening a configurable
appointment.invite form, we directly copy a link to share the
appointment, meaning: a link with only filter_appointment_type_ids
set, with only the selected appointment inside, and with
'all_assigned_resources' for 'resources_choice'.

This simple configuration allows for a quick sharing (on the kanban
view for now). Of course, we only create a new link if we do not find
an existing one with identical config.

Task-4898893

Part-of: odoo/enterprise#88637
Related: odoo/upgrade#8014
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>