[PERF] account_reports: remove join on account.move for aged reports
The Aged Partner reports query in
`_aged_partner_report_custom_engine_common` is joining on the
`account_move` table solely to get the `invoice_date` column. As this
`invoice_date` is aggregated in `ARRAY_AGG(DISTINCT invoice_date)` and
there's already a denormalized field `invoice_date` on
`account_move_line` that is a related stored field, we can drop the join
and use this field instead.

This saves us the need to scan the account.move table when running the
reports query. In a database with 37M amls, 10M account.move (and a
really bad correlation for account_move_line.move_id), this saves us
around 10 minutes for the Aged Receivable report's query.

closes odoo/enterprise#89532

X-original-commit: 2d8ba7c8a76b1332417c83b31eb2a7d49e687402
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Aurélien van Delft (avd) <avd@odoo.com>