[IMP] ai,*: simplify tool declaration with @register_ai_tool
*: ai_calendar,test_ai

Remove the need for server action records when declaring AI tools. Tools
are now defined by decorating methods with @register_ai_tool, making
declarations cleaner and more maintainable.

Schema validators are introduced to support basic parameters schema. These
validators are used at app loading via _register_hook to ensure tool parameters
are properly defined. More complicated schemas will be supported in the
future if necessary.

Input validation is also introduced. Whenever a tool is called, the input is
tested against the schema declared for the called tool, raising a validation
error if the input doesn't match the tool schema.

Important note: It's not enough to decorate a method to register it as an AI
tool. A corresponding `ai.tool` record has to be created. It has to be done this
way to avoid auto-generation of `ai.tool` records. Moreover, `ai.tool` records
have to be created for finer control since the tools are linked to the
`ai.topic` records.

task-4791949

closes odoo/enterprise#85995

Related: odoo/upgrade#7725
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>
Co-authored-by: Joseph (jcb) <jcb@odoo.com>