[FIX] l10n_br_edi: automatically determine other common invoice purposes
We hardcoded the standard purpose (finNFe) and goal for invoices,
credit notes and debit notes. In certain cases however, the purpose is
must be different for EDI to succeed (based on the combination of sold
product types, vendor and customer). In master we will likely add some
fields on the operation type so the user can choose, but in stable we
can already determine some other common cases based on the operation
type.

Using XML ids was considered but since users can create and edit
operation types [1] it's better to use the technical names.

[1] Since odoo/enterprise#73198

opw-4832447

closes odoo/enterprise#89842

X-original-commit: ed0c34493f12e6f2d0e56a016087d466c53afd2d
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>