[FIX] l10n_pe_reports_stock: prevent error in stock move PLE reports
Currently, an error occurs during the valuation of the Stock Move PLE reports.

Steps to Reproduce:

 - Install the l10n_pe_reports_stock module.
 - Create a new company with the country set to Peru, and switch to this
   company.
 - Go to Receipts in Inventory, create a new receipt, and validate it.
 - Go to Valuation and click on PLE Reports.
 - Select the start date as the day after today, and click on PLE 13.1.

KeyError: 'category'

This error occurs during the valuation of the Stock Move PLE reports.
In the query, we retrieve product_template.categ_id as category_id [1],
but later category [2] is used to access category_id. Since the category key
is not present in the line, this causes the error.

[1] https://github.com/odoo/enterprise/blob/95d20fca13efa0977018b0dc260efd410891fce6/l10n_pe_reports_stock/wizard/stock_move_ple_report.py#L241

[2] https://github.com/odoo/enterprise/blob/95d20fca13efa0977018b0dc260efd410891fce6/l10n_pe_reports_stock/wizard/stock_move_ple_report.py#L303

This commit ensures that category_id is used in place of category for
safe evaluation.

sentry-6723761615

closes odoo/enterprise#89505

X-original-commit: 72b4a223ec4e2e09fc93bebd150923abba37a8df
Signed-off-by: Adrien Widart (awt) <awt@odoo.com>
Signed-off-by: Ashutosh Sharma (assh) <assh@odoo.com>