[REF] quality, *: Merge tablet_image and iot_picture widget
This commit was part of the effort of removing code that is no longer
relevant, as well as removing legacy code (QUnit tests).

Because the tablet_image field was not used in a standalone way, it could
be removed and its code moved to the widget that was extending it, iot_picture.

The existing QUnit test has been converted to the HOOT test suite.

task-4028356

closes odoo/enterprise#89373

Signed-off-by: Pierre Paridans (app) <app@odoo.com>