[FIX] pos_appointment: load missing appointment_svg template
Steps to reproduce:
-------------------------
- Open pos enabled with bookings.
- Open booking tab and in gantt view remove groupby filter.

Issue:
-------
There will be a tb.

Cause:
--------
When there are no bookings available it tries to load no booking svg template,
which is not loaded in the assets.

Fix
---
We have loaded the missing asset file to ensure the SVG displays correctly
when there are no bookings.

closes odoo/enterprise#89901

Task: 4912389
X-original-commit: 30aedcb363b7710f9c0b2de2e54684754de4c87d
Signed-off-by: Stéphane Vanmeerhaeghe (stva) <stva@odoo.com>
Signed-off-by: Jainam Shah (sjai) <sjai@odoo.com>