[IMP] l10n_mx_edi*: Refactor the CFDI generation to always be accepted by the SAT
Simplify the CFDI generation. Now the totals are always computed from the lines.
That way, it's impossible for the SAT to refuse a CFDI since it's built in the
way the SAT checks the validity of the file.

The difference between the CFDI total and the original documents is put using the discount if available.
If not, we add a fake line in the CFDI.

closes odoo/enterprise#94602

Task-id: 4926200
X-original-commit: 911732fbfc8666e3c4d94e830e255eeb9635516b
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Laurent Smet (las) <las@odoo.com>