[FIX] l10n_es_reports: exclude current period credit notes BOE 349
On a **l10n_es** fiscal position:

- Create a bill for the current period and then create a credit note for this bill.
- In the tax report, under model 349, download the BOE report.

Since the bill has been fully canceled within the same period, no rectification line should appear in the BOE report (correct behavior).

- Create a bill from a previous period.
- Create a credit note for this bill in the current period.
- Download the BOE report from model 349.

A rectification line appears for the previous period bill, which is expected.
However, a rectification line also appears for the current period bill, which is incorrect, as the bill was canceled within the same period.

This commit prevents moves from the current period from being included as rectification lines in the BOE report.

opw-4895636

closes odoo/enterprise#90049

X-original-commit: 564c7cb97199d52b7f2c3d47e6586334d9c864c2
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>
Signed-off-by: Guillaume Teboul-Tornezy (gute) <gute@odoo.com>