[FIX] marketing_automation_website_sale: correct domain for marketing automation
We had a traceback when trying to create an Anniversary Discount marketing campaing
due to the domain's construction.

Steps to reproduce:
-------------------
* Go to Marketing Automation app
* Go to eCommerce tab --> anniversary discount
* Click " Create Campaign"

> Observation:
  File "/data/build/enterprise/marketing_automation/models/marketing_activity.py", line 142, in _compute_inherited_domain
    literal_eval(activity.campaign_id.domain or '[]')])

Why the fix:
------------
The problematic line was introduced in this REV: 1aa05c89f3371981922e3dc52d8334104aa425b9.
The domain was originally built in this IMP: cf2124f4f20a805077502587f9a8be0af0a753c3.
The domain was intended to be used dynamically, but after the revision, that usage changed—so the domain needed to be rebuilt.
We've simplified the domain to eliminate overlapping conditions, ensuring the same logic and result, but with improved readability.

opw-4857722

closes odoo/enterprise#90136

X-original-commit: b736247d001ac3f7aac1292fed85faa365c94282
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Guillaume Vardakis (gvar) <gvar@odoo.com>