[IMP] ai(_fields): show message when field cannot be computed
Purpose:
--------
Sometimes, the LLM returns a falsy value because the value to assign to
the field is a falsy value (a boolean, an empty M2X,...).
Currently, an error notification is shown in that case ("Oops, ...").

Changes in this commit allows to show this notification only when the
LLM can't assign a value because some context is missing and it can
confidently enough generate a value. When that's the case, we display
an error message generated by the LLM instead of showing a generic error
message (such as "No financial data about xxx was found").

It also shows more user friendly error messages when a request times out,
or when the connection failed. To do so, the `_request` method of the
`llm_api_service` now re-raises the request exception, so that it can be
re-caught later.

Task-4762422

Part-of: odoo/enterprise#84788
Related: odoo/odoo#217078
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>