[FIX] planning: add tests for the planning calendar view
This commit's purpose is to add some testing for the task-4485911. This
task implemented the use of the new framework tool that allows to create
and delete records in batch in the calendar view. Some js customisation
was needed in order to better fit our use case.

Due to a short timing the testing is done in this followup commit
instead of the initial task.

task-4891810
target : saas-18.4

closes odoo/enterprise#90500

X-original-commit: 2f5241ffc22bee425fc1f4db8c11b4eddfeb0d73
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>