[FIX] documents: remove temporary card on cancel
Before this commit, when aborting a drag & drop sequence in documents
views, the temporary card "ghost" was not removed from the DOM, and
stayed in place.

This was most noticable in tests, where multiple cards would pile up
on the screen.

This commit, adds a cleanup for each added card, to ensure that they are
properly disposed of.

closes odoo/enterprise#89479

Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>
Signed-off-by: Julien Mougenot (jum) <jum@odoo.com>