[FIX] l10n_se_sie4_import: Prevent crash when previous year information is not present in SIE4 file
**Issue**:
Importing a SIE4 file without previous year information causes a traceback.

**Steps to reproduce**:
- Go to Accounting > Settings > Import
- Import SIE 4 file
- Check the box "Import account opening balances"
- Select the right xml and observe the traceback

**Cause**:
The method `_prepare_sie4_opening_balance_move` tries to directly access the previous year:
https://github.com/odoo-dev/enterprise/blob/6d4919658650a006c73d4aaf1f500d67723dda0d/l10n_se_sie4_import/wizard/import_wizard.py#L376C9-L376C58
This results in a traceback when the previous year is not present.

**Solution**:
Make `_prepare_sie4_opening_balance_move` more permissive by falling back to the day before the first day of the current year if the `-1` section is not there.

**Additional Notes**:
The client file does not support UTF8 format, retry with the ISO-8859-1 format in case of UnicodeDecodeError.

opw-4894495

closes odoo/enterprise#92804

X-original-commit: c2627310a948a4868207515aefd569996a864140
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Maxime Noirhomme (noma) <noma@odoo.com>