[FIX] delivery_sendcloud: avoid crash on manual tracking ref
Steps to reproduce:

1. Install delivery_sendcloud.
2. Create a new shipping method (Inventory -> Configuration)
   - Provider: sendcloud
   - Put the public and secret key
   - Assign Delivery Product
3. Create a sales order with a deliverable product
4. Confirm the sale order
5. Go to the delivery
6. Manually enter a random tracking reference.
7. Make sure carrier is set to the Sendcloud delivery method
8. Click on "tracking," or get back to the sales order (click "preview").
9. A traceback occurs.

Issue:

A TypeError: 'bool' object is not subscriptable

when accessing sendcloud_parcel_ref, which is expected to be a list or JSON but
is instead a boolean.

Causes:

The code assumes the presence of valid Sendcloud parcel reference data, but
manually entering a tracking reference bypasses the _send_shipment method that
populates this field. As a result, sendcloud_parcel_ref can remain False.

Solution:

Return None if there is no Sendcloud parcel reference, which will
raise a UserError from the below line
https://github.com/odoo/odoo/blob/f08981c5897297bf37f91225a2523b038b09d0d6/addons/stock_delivery/models/stock_picking.py#L287-L288

Co-Author By - alsh@odoo.com

opw-4894677

closes odoo/enterprise#91572

X-original-commit: 212f0e0a788de35a3a9f17c14798ac1e32b4601a
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>