[IMP] accountant_knowledge, sign: add signature plugin
The /signature command previously inserted the user's email signature
into documents. However, because the email signature is already appended
automatically when sending messages through the mail composer - and
doesn't serve much purpose outside of email communication - it was
removed from the list of available commands.

In this commit, we're reintroducing the /signature command in the
accountant_knowledge module, but with a new purpose. Instead of adding
an email signature, the updated command will allow users to insert an
electronic signature directly into their reports.

Task-4840940

Part-of: odoo/enterprise#87928
Related: odoo/odoo#214712
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>