[FIX] mrp_workorder: access right error for mrp users
**PROBLEM**
If a user is in the mrp.group_mrp_user group, but does not belong to hr.group_hr_user, he can't access the shop floor app.

**STEP TO REPRODUCE**
1. connect with a user which is a user of manufactring, but not a user of hr.
2. try to go on the shop floor app and notice there is an access right error.

**CAUSE**
When connecting to the shop floor app, we are trying to get the barcode field on all employee (because we need them  if we want to identify an employee on the shop floor app using their barcode).
This was added in this commit: https://github.com/odoo/enterprise/commit/b3fb0073a15adcc799a5681284f0cfd2308764b8
The barcode field is only accessible to member of hr.group_hr_user.

**FIX**
Instead of getting the barcode of all employee using  `get_all_employee()`, we do a rpc call to query the employee the barcode belong to.

opw-4905206

closes odoo/enterprise#91555

X-original-commit: a4e909a64f4239e1f5a17be337e4df2e952a3e12
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Léo Gizard (legi) <legi@odoo.com>